Use serde instead of rustc-serialize
rustc-serialize is unable to decode generic json values because
the Json enum does not implement the Decodable trait. This is needed
to be able to decode certain messages in the webdriver spec that
return arbitrary json types. Replaced rustc-serialize with serde_json,
the serde Value type is reexported as JsonValue.

Added a couple functions that use element references, i.e. Json
objects:

- a switch_to_frame() method to switch contexts
- a FrameContext type to help switching context and back