Add execute command to the CLI
Usabilitywise this is extremely experimental, it adds an execute
command to the CLI to run javascript code. It still requires you
to build a proper command that returns a value e.g.

    >> execute return 42;

Which is not very convenient. But I have not improved on it yet
because there a number of other issues around this command, namely
proper print out of the return value. Right now string values are
printed correctly but others are not - they are json strings with
the json escaping in place.