Do not drop an invalid session
On drop DriverSession deletes the existing session by default. However
inside the create_session method an invalid session was being created
(with an empty session id string) and on failure it was being dropped.
The attempt to remove a non existing session would always fail.

We can avoid the meaningless API call on drop by avoiding to build the
DriverSession type until we are sure a session id is available.