Workaround non-conformant chrome element ref
The switch frame command failed to execute in chrome, the error message
is pretty explicit it requires the payload of the method to carry the
`ELEMENT` attribute instead of the attribute from the w3c spec
(element-6066-11e4-a52e-4f735466cecf).

I think the upstream issue is https://bugs.chromium.org/p/chromedriver/issues/detail?id=1992

The proposed workaround for this is to also serialize the
ElementReference with an ELEMENT attribute. There is the risk of
clashing with other implementations but so far have I have encoutered no
issues.

Removed the serialization test from the messages module.