Run tests with chromedriver
Add travis settings to run tests with the stable version of chrome, a
script to download chromedriver is available in
bin/download_chromedriver. To avoid problems with the travis container
the tests run in headless mode without the chrome sandbox:

    https://github.com/travis-ci/travis-ci/issues/7150#issuecomment-359145343

To make the tests generic over different browsers they were wrapped in a
macro whose purpose is to generate test modules for different drivers
(currently chrome and firefox). At runtime tests can determine the
driver they are using with the function test_browser().  It returns an
enum that identifies the type of driver in use and provides constructors
for the driver or session.

Runtime checks for the browser being executed either use this enum or
the browser name.

Some loose ends are left unaddressed

- we can skip a test at runtime but there is no way to indicate this in
the test output, one just returns
- part of the `execute` test is being skipped for chrome, the error
strings and types are very different from firefox
- Tests using the property method skipped in chrome because this is not
implemented https://bugs.chromium.org/p/chromedriver/issues/detail?id=1936